utils methods

+++ src/main/java/com/uz/justplan/lookup/PriorityRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.ListPagingAndSortingRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface PriorityRepository extends CrudRepository<Priority, Integer>, ListPagingAndSortingRepository<Priority, Integer> {
+public interface PriorityRepository extends CrudRepository<Priority, Long>, ListPagingAndSortingRepository<Priority, Long> {
     Optional<Priority> findByCompanyIdAndNameIgnoreCase(long companyId, String name);
 
     Page<Priority> findByCompanyIdAndActive(long companyId, boolean active, Pageable pageable);
